pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "fastapi>=0.116.1",
    "levenshtein>=0.27.1",
    "numpy>=2.3.2",
    "openai>=1.97.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.42",
    "tenacity>=9.1.2",
    "tweepy>=4.16.0",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
]
